backend/pyproject.toml
[project]
name = "PaceUp"
version = "0.1.0"
description = "PaceUp is a platform for tracking your running activities and analyzing your performance."
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.116.1",
    "httpx>=0.28.1",
    "pydantic-settings>=2.10.1",
    "pytest>=8.4.1",
    "python-dotenv>=1.1.1",
    "sqlalchemy>=2.0.42",
    "uvicorn>=0.35.0",
    "psycopg2-binary>=2.9.9",
    "requests>=2.31.0",
    "openai>=1.0.0",
    "redis>=5.0.0",
]

[tool.pytest.ini_options]
pythonpath = "."